requirements.txt
python-telegram-bot==13.7
sqlalchemy==1.4.22
python-dotenv==0.19.1